wip verify using spies
