+++ src/suppose/verify.ts
@@ -101,76 +101,6 @@ ${failedSuppositions
   }
 }
 
-/**
- *
- * This function should either: use the spy, or build upon the verify & suppositions system.
- * I don't know which one is better yet.
- */
-
-// export function verifyThat<p>(
-//   mock: (param: p) => any,
-// ): { wasCalledWith: (param: p) => void; wasCalled: () => void };
-// export function verifyThat<p1, p2>(
-//   mock: (param1: p1, param2: p2) => any,
-// ): {
-//   wasCalledWith: (param1: p1, param2: p2) => void;
-//   wasCalled: () => void;
-// };
-// Make it work for any number of params
-export function verifyThat<Params extends any[], Result>(
-  mock: (...args: Params) => Result
-): {
-  wasCalledWith: (...params: Params) => void;
-  wasCalled: () => void;
-};
-export function verifyThat<Params extends any[], Result>(
-  mockedFunction: (...args: Params) => Result
-) {
-  return {
-    wasCalledWith: (
-      ...params: [
-        Params[0],
-        Params[1],
-        Params[2] extends {} ? Params[2] : never,
-        Params[3] extends {} ? Params[3] : never,
-        Params[4] extends {} ? Params[4] : never
-      ]
-    ) => {},
-  };
-}
-
-function mockedFunction(a: string, number: number) {
-  return "a";
-}
-
-function test2(
-  a: string,
-  b: number,
-  c: string,
-  d: number,
-  e: string,
-  f: number,
-  z: { x: 1 }
-) {
-  return "a";
-}
-
-function test3(a: { x: number; fullName: string }, email: string) {
-  return "a";
-}
-
-verifyThat(mockedFunction).wasCalledWith("azezae", 1);
-
-verifyThat(test2).wasCalledWith("azea", 1, "azeaz", 2, "azeae", 3, { x: 1 });
-
-verifyThat(test3).wasCalledWith(
-  {
-    fullName: "azeaze",
-    x: 1,
-  },
-  "victor"
-);
-
 function parseSuppositionText(supp: SuppositionCount) {
   if (supp === "NEVER") {
     return "\nIt was expected to never be called";
